[Stats]Fix overflow issues. (#1068)
* Support int64 in CDN Downloads aggregate.
* Update Gallery Core package.
* Missed a long.

+++ src/Stats.AggregateCdnDownloadsInGallery/AggregateCdnDownloadsJob.cs
@@ -45,7 +45,7 @@ namespace Stats.AggregateCdnDownloadsInGallery
             (
                 [PackageRegistrationKey]    INT             NOT NULL,
                 [PackageVersion]            NVARCHAR(255)   NOT NULL,
-                [DownloadCount]             INT             NOT NULL,
+                [DownloadCount]             BIGINT          NOT NULL,
             )";
 
         private const string _updateFromTempTable = @"
